Better packages support (#67)
* Move rules function API from rules field to each rule usage

It is more flexible for each rule to be able to be configured as a
function rather than relying on the `rules` field itself to be one. In
particular, this makes it easier to use third party shareable configs
that rely on rule functions for package-level rules.

This additionally moves the rules function processing to a
pre-processing step so that the resulting rule usages can be validated.